Delete logo_ecopetrol.jpg
o the attention of GITHUB ,  

  

Telefónica Cybersecurity & Cloud Tech manages the fraudulent actions against  ECOPETROL and all issues related to security incidents against this company.  

  

We have detected a trademark infringement incident against our client , on a website hosted by GitHub , in the following URL(s):   

  

  https://github.com/diacosta/ecopetrol-reto 

 https://github.com/diacosta/ecopetrol-reto/blob/master/image_app/static/logo_ecopetrol.jpg?raw=true

https://raw.githubusercontent.com/diacosta/ecopetrol-reto/master/image_app/static/logo_ecopetrol.jpg
  
https://github.com/diacosta/ecopetrol-reto/blob/b68f5730c99929e24d20dd0306dbde5d9e1ed79e/image_app/static/s.jpg

This website uses our client's registered trademark and logo in order to act in behalf of them,  without any authorization from our client and without including any disclaimer to clarify they do not have any relationship with our client, which supposes an impersonation of our client in order to deceive users and make them believe  it is a legitimate page. For this reason, this content represents a misuse of the industrial property of Ecopetrol .  

  

In the following official link/s, you can check the registration information of our client's trademark/s:  

  

https://www3.wipo.int/branddb/en/showData.jsp?ID=EMTM.008128373   

  

We also have attached  the representation sheet that authorizes us to act on behalf of our client.  

  

We need your collaboration to resolve this incident, by requesting your client to clarify the purpose of this trademark infringement, confirming the authorization for its use, or removing the infringing content from the website. If you need more information regarding this incident, please contact our SOC 24/7 at +34 900 102 230 (option 7) or by replying to this email.  

  

Thank you very much for your attention. Looking forward to your reply.  

  

Regards,